+++ tests/test_wood_heat_capacity.py
@@ -1,0 +1,11 @@
+"""
+Tests for the wood heat capacity module.
+"""
+
+import chemics as cm
+from pytest import approx
+
+
+def test_cp_wood():
+    cp = cm.cp_wood(12, 340)
+    assert cp == approx(1.91, rel=1e-2)